hml-metadata-generator-package/requirements.txt
pyyaml>=6.0
anthropic>=0.18.0
